Allow installing on newer Python versions
Although it has no effect, this package should be able to be installed on newer python versions. The [PyPi package](https://pypi.org/project/typing/) is able to be and explicitly lists this behavior:

> NOTE: in Python 3.5 and later, the typing module lives in the stdlib, and installing this package has NO EFFECT.

This is helpful so that if other packages rely on this package (like `typing_extensions`) they can still resolve it on a later python version.